add normal JHUGen, DM variation now marked with DM
