+++ bin/JHUGen/runcmsgrid_template_DM.sh
@@ -9,9 +9,12 @@ echo "%MSG-MG5 random seed used for the run = $rnum"
 ncpu=${3}
 echo "%MSG-MG5 number of cpus = $ncpu"
 
-eval `scramv1 runtime -sh`
-
-cd BASEDIR/
+#eval `scramv1 runtime -sh`
+cd BASEDIR/pdfs
+wget http://pcteserver.mi.infn.it/~nnpdf/nnpdf30/NNPDF30_lo_as_0130.LHgrid.tgz
+tar -zxvf ./NNPDF30_lo_as_0130.LHgrid.tgz
+rm ./NNPDF30_lo_as_0130.LHgrid.tgz
+cd ..
 
 GENCOMMAND
 file=Out.lhe
@@ -21,8 +24,14 @@ tail=`wc -l $file | awk -v tmp="$head" '{print $1-2-tmp}'`
 
 tail -${tail} $file                                     >  ${file}_tail_dec
 sed "s@\*\*\*@1.000@g"     ${file}_tail_dec              > ${file}_tail
+./boltdmdec ${file}_tail
+sed "s@-1000022@1000022@g" ${file}_tail_dec             >  ${file}_tail
+head -$head $file                                       > cmsgrid.lhe
+echo "  "XSECTION"   "XSECUNC"  1.00000000000E-00 10001" >> cmsgrid.lhe
+echo "</init>"                                         >> cmsgrid.lhe
 cat ${file}_tail                                       >> cmsgrid.lhe
 rm ${file}
 rm ${file}_tail
+rm ${file}_tail_dec
 mv cmsgrid.lhe ../cmsgrid_final.lhe
 cd ..
